# api/openapi.yaml
openapi: 3.1.0
info:
  title: Qoder Terminal User API
  version: 0.1.0
  description: 登录鉴权与用户行为记录。data / analyst 通过 JWKS 验证本服务签发的 JWT。
servers:
  - url: http://localhost:8084
paths:
  /health:
    get:
      operationId: getHealth
      summary: Get health
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
  /.well-known/jwks.json:
    get:
      operationId: getJwks
      summary: Get JWKS
      description: RS256 公钥集合，供其他服务验签
      security: []
      responses:
        "200":
          description: JWK Set（仅公钥）
          content:
            application/json:
              schema: { type: object, required: [keys], properties: { keys: { type: array, items: { type: object } } } }
  /v1/auth/register:
    post:
      operationId: register
      summary: Register with invite code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: 注册成功并返回令牌
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/auth/login:
    post:
      operationId: login
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/me:
    get:
      operationId: getMe
      summary: Get current user
      responses:
        "200":
          description: 当前用户
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/activities:
    get:
      operationId: listActivities
      summary: List my activities
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: 按时间倒序
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Activity" } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
    post:
      operationId: recordActivity
      summary: Record an activity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActivityRequest" }
      responses:
        "201":
          description: 已记录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
    RegisterRequest:
      type: object
      required: [inviteCode, username, password, displayName]
      properties:
        inviteCode: { type: string }
        username: { type: string, pattern: "^[a-z0-9_]{3,32}$" }
        password: { type: string, minLength: 8, maxLength: 72 }
        displayName: { type: string, maxLength: 100 }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    TokenResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken: { type: string, description: "RS256 JWT；iss=qoder-terminal-user，aud=qoder-terminal，sub=用户 ID" }
        tokenType: { type: string, enum: [Bearer] }
        expiresIn: { type: integer, description: 秒 }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      required: [id, username, displayName, role, createdAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [USER, ADMIN] }
        createdAt: { type: string, format: date-time }
    ActivityType:
      type: string
      enum: [COMMAND, ASK, PANEL_OPEN]
    ActivityRequest:
      type: object
      required: [type, content]
      properties:
        type: { $ref: "#/components/schemas/ActivityType" }
        content: { type: string, minLength: 1, maxLength: 2000, example: "700 Q" }
    Activity:
      type: object
      required: [id, type, content, createdAt]
      properties:
        id: { type: integer, format: int64 }
        type: { $ref: "#/components/schemas/ActivityType" }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }